Refactoring Content class, Freemarker support #109
