+++ chat-workflow-spring-boot-starter/src/main/java/org/finos/symphony/toolkit/workflow/sources/symphony/handlers/freemarker/AbstractClassFieldConverter.java
@@ -1,0 +1,20 @@
+package org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker;
+
+import java.lang.reflect.Field;
+
+public abstract class AbstractClassFieldConverter extends AbstractFieldConverter {
+
+	private Class<?> forClass;
+
+	public AbstractClassFieldConverter(int priority, Class<?> forClass) {
+		super(priority);
+		this.forClass = forClass;
+	}
+
+	@Override
+	public boolean canConvert(Field f) {
+		return forClass.isAssignableFrom(f.getType());
+	}
+	
+	
+}
