+++ chat-workflow-spring-boot-starter/src/main/java/org/finos/symphony/toolkit/workflow/sources/symphony/handlers/freemarker/FieldConverter.java
@@ -1,0 +1,11 @@
+package org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker;
+
+import java.lang.reflect.Field;
+
+public interface FieldConverter extends WithField {
+
+	public int getPriority();
+	
+	public boolean canConvert(Field f);
+		
+}
