+++ chat-workflow-spring-boot-starter/src/main/java/org/finos/symphony/toolkit/workflow/sources/symphony/handlers/freemarker/Mode.java
@@ -1,0 +1,5 @@
+package org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker;
+
+enum Mode {
+	DISPLAY, DISPLAY_WITH_BUTTONS, FORM
+}
