+++ chat-workflow-spring-boot-starter/src/test/java/org/finos/symphony/toolkit/workflow/AbstractMockSymphonyTest.java
@@ -6,6 +6,19 @@ import org.finos.symphony.toolkit.workflow.java.perform.PerformerConfig;
 import org.finos.symphony.toolkit.workflow.java.resolvers.ResolverConfig;
 import org.finos.symphony.toolkit.workflow.sources.symphony.SymphonyBot;
 import org.finos.symphony.toolkit.workflow.sources.symphony.SymphonyWorkflowConfig;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.AuthorConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.BeanConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.BooleanConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.CashTagConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.CollectionConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.EnumConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.HashTagConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.IDConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.InstantConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.NumberConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.RoomConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.StringConverter;
+import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.freemarker.UserConverter;
 import org.finos.symphony.toolkit.workflow.sources.symphony.handlers.jersey.JerseyAttachmentHandlerConfig;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
@@ -27,6 +40,19 @@ import com.symphony.api.pod.UsersApi;
 @ExtendWith(SpringExtension.class)
 
 @SpringBootTest(classes = { 
+		AuthorConverter.class,
+		BeanConverter.class,
+		BooleanConverter.class, 
+		CashTagConverter.class,
+		CollectionConverter.class,
+		EnumConverter.class,
+		HashTagConverter.class,
+		IDConverter.class,
+		InstantConverter.class,
+		NumberConverter.class,
+		RoomConverter.class,
+		StringConverter.class,
+		UserConverter.class,
 		PerformerConfig.class,
 		ResolverConfig.class,
 		TestWorkflowConfig.class,
